backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
supabase
transformers
torch
scikit-learn
numpy